+++ CHANGELOG.md
@@ -9,7 +9,10 @@
 ### Features ⭐
 
 - Add registered patients line-list report
-- Generic Vaccine line list report
+- Add India assistive technology device line list report
+- Add Iraq assistive technology device line list report
+- Add PNG assistive technology device line list report
+- Add Generic Vaccine line list report
 - Allow disabling reports through server config
 
 ## Tweaks ⚖️